% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{add_covariates}
\alias{add_covariates}
\title{Add several covariates to movement observations
\code{add_covariates} add several covariates to a data frame with movement
information. It adds : distance between location, spatial angle, speed,
smoothed speed, persistence and rotation velocity (calculated with spatial
angle).}
\usage{
add_covariates(x, coord.names = c("x", "y"), smoothed = F,
  timecol = "dateTime", units = "hour", radius = NULL)
}
\arguments{
\item{x}{data.frame with locations}

\item{coord.names}{names of coordinates column in \code{x}}

\item{smoothed}{whether speed are smoothed or not}

\item{timecol}{names of POSIXct time column}

\item{units}{units for time calculation. Default "hour"}

\item{radius}{for spatial angle calculations}
}
\value{
data.frame increased with
}
\description{
Add several covariates to movement observations
\code{add_covariates} add several covariates to a data frame with movement
information. It adds : distance between location, spatial angle, speed,
smoothed speed, persistence and rotation velocity (calculated with spatial
angle).
}
\examples{
calc_dist(df,coord.names = c("x","y"), smoothed = T)
}
\author{
Rémi Patin, \email{remi.patin@normale.fr}
}
