% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{angular_speed}
\alias{angular_speed}
\title{Calculate angular speed along a trajectory}
\usage{
angular_speed(x, coord.names = c("x", "y"))
}
\arguments{
\item{x}{data.frame with locations}

\item{coord.names}{names of coordinates column in \code{x}}
}
\value{
vector of turning angle.
}
\description{
\code{spatial_angle} calculate turning angle between locations, taking a
dataframe as input.
}
\examples{
\dontrun{calc_speed(df,coord.names = c("x","y"), timecol = "dateTime", smoothed = T)}
}
\author{
Remi Patin, Simon Benhamou.
}
