% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_rowSums}
\alias{apply_rowSums}
\title{apply_rowSums}
\usage{
apply_rowSums(rupt, x)
}
\arguments{
\item{rupt}{current estimated breaks in signal}

\item{x}{bivariate signal}
}
\description{
Internal function for Expectation-Maximization (EM) algorithm.
}
