% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_diag.var}
\alias{argcheck_diag.var}
\title{Check for argument 'diag.var'}
\usage{
argcheck_diag.var(diag.var, seg.var)
}
\arguments{
\item{diag.var}{names of the variables on which
statistics are calculated.}

\item{seg.var}{for behavioral segmentation: names of the variables used for
segmentation (either one or two names).}
}
\value{
a vector of character string
}
\description{
Check whether argument 'diag.var' was provided.
If not, propose default value.
}
