% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_Kmax}
\alias{argcheck_Kmax}
\title{Check for argument 'Kmax'}
\usage{
argcheck_Kmax(Kmax, lmin, datalength)
}
\arguments{
\item{Kmax}{maximum number of segments.}

\item{lmin}{minimum length of segments.}

\item{datalength}{length of data provided}
}
\value{
an integer
}
\description{
Check whether argument 'Kmax' was provided and is adequate
before subsampling. Propose adequate value if Kmax is not
provided.
}
