% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_lmin}
\alias{argcheck_lmin}
\title{Check for argument 'lmin'}
\usage{
argcheck_lmin(lmin, is_segclust)
}
\arguments{
\item{lmin}{minimum length of segments.}

\item{is_segclust}{TRUE if function is called from \code{\link{segclust}} ; 
FALSE otherwise, if function is called from \code{\link{segmentation}}.}
}
\value{
a NULL object
}
\description{
Check whether argument 'lmin' was provided and is adequate
before subsampling
}
