% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_ncluster}
\alias{argcheck_ncluster}
\title{Check for argument 'ncluster'}
\usage{
argcheck_ncluster(ncluster, Kmax)
}
\arguments{
\item{ncluster}{number of cluster into which segments should be grouped. Can
be a vector if one want to test several number of clusters.}

\item{Kmax}{maximum number of segments.}
}
\value{
a NULL object
}
\description{
Check whether argument 'ncluster' was provided and is adequate
}
