% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_order.var}
\alias{argcheck_order.var}
\title{Check for argument 'order.var'}
\usage{
argcheck_order.var(order.var, diag.var)
}
\arguments{
\item{order.var}{names of the variable with which states are ordered.}

\item{diag.var}{names of the variables on which
statistics are calculated.}
}
\value{
a vector of character string
}
\description{
Check whether argument 'order.var' was provided.
If not, propose default value.
}
