% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_ordering}
\alias{argcheck_ordering}
\title{Check for argument 'order'}
\usage{
argcheck_ordering(order, seg.type)
}
\arguments{
\item{order}{TRUE or FALSE depending on whether cluster be ordered}

\item{seg.type}{types of the segmentation.}
}
\value{
a boolean
}
\description{
Check whether argument 'order' was provided for plot.segmentation
and segmap. If not, propose default value.
}
