% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_scale.variable}
\alias{argcheck_scale.variable}
\title{Check for argument 'scale.variable'}
\usage{
argcheck_scale.variable(scale.variable, is_segclust)
}
\arguments{
\item{scale.variable}{minimum length of segments.}

\item{is_segclust}{TRUE if function is called from \code{\link{segclust}} ; 
FALSE otherwise, if function is called from \code{\link{segmentation}}.}
}
\value{
a boolean
}
\description{
Check whether argument 'scale.variable' was provided.
If not, propose default value.
}
