% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_seg.var}
\alias{argcheck_seg.var}
\title{Check for argument 'seg.var'}
\usage{
argcheck_seg.var(x, seg.var, is_segclust)
}
\arguments{
\item{x}{data used for segmentation. Supported: data.frame, Move object,
ltraj object}

\item{seg.var}{for behavioral segmentation: names of the variables used for
segmentation (either one or two names).}

\item{is_segclust}{TRUE if function is called from \code{\link{segclust}} ; 
FALSE otherwise, if function is called from \code{\link{segmentation}}.}
}
\value{
a list with a data.frame and a vector with two character strings
}
\description{
Check whether argument 'seg.var' was adequately provided.
If provided, also check for its length (1 or 2) and
for the existence of corresponding column names in x
If unprovided, use default value (segmentation only) and tests
if column names exists.
}
