% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_segclust}
\alias{argcheck_segclust}
\title{Check for argument 'ncluster' and 'nseg'}
\usage{
argcheck_segclust(ncluster, nseg, ncluster.BIC, Kopt.BIC)
}
\arguments{
\item{ncluster}{number of cluster}

\item{nseg}{number of segment}

\item{ncluster.BIC}{optimal number of cluster selected by BIC}

\item{Kopt.BIC}{optimal number of segment selected by BIC for 
each number of cluster}
}
\value{
a list with two integers
}
\description{
Check whether argument 'ncluster' and 'nseg' were provided.
 If not, propose default value based on BIC.
}
