% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_segmentation}
\alias{argcheck_segmentation}
\title{Check for argument 'nseg'}
\usage{
argcheck_segmentation(nseg, Kopt.lavielle)
}
\arguments{
\item{nseg}{number of segment}

\item{Kopt.lavielle}{optimal number of segment selected with 
Lavielle's criterium}
}
\value{
an integer
}
\description{
Check whether argument  'nseg' was provided.
 If not, propose default value based on Lavielle's criterium
}
