% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_checks.R
\name{argcheck_type_coord}
\alias{argcheck_type_coord}
\title{Check for deprecated 'type' and 'coord.names' argument}
\usage{
argcheck_type_coord(...)
}
\arguments{
\item{...}{additional parameters transmitted from \code{\link{segmentation}}
or \code{\link{segclust}}}
}
\value{
a NULL object
}
\description{
Check whether argument 'type' and 'coord.names' 
were provided and communicate adequately if need be.
}
