% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_repmat}
\alias{arma_repmat}
\title{arma_repmat}
\usage{
arma_repmat(A, n, m)
}
\arguments{
\item{A}{matrix}

\item{n}{number of repetition in line}

\item{m}{number of repetition in column}
}
\description{
C++ Armadillo version for repmat function. Repeat a matrix in bloc.
}
