% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{calc_BIC}
\alias{calc_BIC}
\title{Calculate BIC}
\usage{
calc_BIC(likelihood, ncluster, nseg, n)
}
\arguments{
\item{likelihood}{log-likelihood}

\item{ncluster}{number of cluster}

\item{nseg}{number of segment}

\item{n}{number of observations}
}
\value{
a data.frame with BIC, number of cluster and number of segment
}
\description{
\code{BIC} calculates BIC given log-likelihood, number of segment and number of class
}
\examples{
\dontrun{calc_stat_states(data, diag.var=c("dist","angle"),
order.var='dist',type='hmm',hmm.model=mod1.hmm)}
}
