% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distance between locations}
\usage{
calc_dist(x, coord.names = c("x", "y"), smoothed = F)
}
\arguments{
\item{x}{data.frame with locations}

\item{coord.names}{names of coordinates column in \code{x}}

\item{smoothed}{whether distance are smoothed or not}
}
\value{
vector of distance
}
\description{
\code{calc_dist} calculate distance between locations, taking a dataframe as input. Distance can also be smoothed over the two steps before and after the each point.
}
\examples{
\dontrun{calc_dist(df,coord.names = c("x","y"), smoothed = T)}
}
\author{
Remi Patin
}
