% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{calc_stat_states}
\alias{calc_stat_states}
\title{Calculate state statistics}
\usage{
calc_stat_states(data, df.segm, diag.var, order.var = NULL)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{df.segm}{output of prep_segm function}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}
}
\value{
a data.frame with mean and variance of the different states
}
\description{
\code{calc_stat_states} calculates statistics of a given segmentation : mean
and variance of the different states.
}
\examples{
\dontrun{calc_stat_states(data, diag.var = c("dist","angle"),
order.var='dist', type='hmm',hmm.model=mod1.hmm)}
}
