% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Kmax.R
\name{chooseseg_lavielle}
\alias{chooseseg_lavielle}
\title{Internal Function for choosing optimal number of segment}
\usage{
chooseseg_lavielle(J, S = 0.75, ...)
}
\arguments{
\item{J}{likelihood for each number of segment}

\item{S}{threshold for choosing the number of segment. See
adehabitatLT::chooseseg}

\item{...}{additional arguments}
}
\value{
a list with optimal number of segment and full data.frame of the
  calculus
}
\description{
Choosing optimal number of segment using Marc Lavielle's method. From Emilie
Lebarbier. Method based on identifying breaks in the slope of the contrast.
}
