% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colsums_sapply}
\alias{colsums_sapply}
\title{colsums_sapply}
\usage{
colsums_sapply(i, rupt, x, mu, tau)
}
\arguments{
\item{i}{number of signal}

\item{rupt}{current estimated breaks in signal}

\item{x}{bivariate signal}

\item{mu}{mean parameter for each signal}

\item{tau}{tau}
}
\description{
Internal function for Expectation-Maximization (EM) algorithm.
}
