% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DynProg_algo_cpp}
\alias{DynProg_algo_cpp}
\title{DynProg_algo_cpp}
\usage{
DynProg_algo_cpp(matD, Kmax)
}
\arguments{
\item{matD}{Cost Matrix}

\item{Kmax}{number of segments}
}
\description{
This function finds the best segmentation given a Cost Matrix using a
dynamic programming algorithm. C++ implementation of \link{DynProg}
}
