% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_EM.algo_simultanee.R
\name{EM.algo_simultanee}
\alias{EM.algo_simultanee}
\title{EM.algo_simultanee calculates the MLE of phi for given change-point instants}
\usage{
EM.algo_simultanee(x, rupt, P, phi, eps = 1e-06, sameSigma = FALSE)
}
\arguments{
\item{x}{bivariate signal}

\item{rupt}{the sequence of change points}

\item{P}{number of clusters}

\item{phi}{starting value for the  parameter}

\item{eps}{eps}

\item{sameSigma}{TRUE if segments have the same variance}
}
\value{
a list with  phi, the MLE, tau =(taukj) the probability for segment k
  to belong to class,lvinc = lvinc,empty = empty,dv = dv
}
\description{
EM.algo_simultanee calculates the MLE of phi for given change-point instants
}
