% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegTraj_EM_cpp.R
\name{EM.algo_simultanee_Cpp}
\alias{EM.algo_simultanee_Cpp}
\title{EM.algo_simultanee caculates the MLE of phi for given change-point instants
 and for a fixed number of clusters
 @rdname EM.algo_simultanee}
\usage{
EM.algo_simultanee_Cpp(x, rupt, P, phi, eps = 1e-06, sameSigma = FALSE)
}
\description{
EM.algo_simultanee caculates the MLE of phi for given change-point instants
 and for a fixed number of clusters
 @rdname EM.algo_simultanee
}
