% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gmixt_algo_cpp}
\alias{Gmixt_algo_cpp}
\title{Gmixt_algo_cpp}
\usage{
Gmixt_algo_cpp(zi, lgi, P, mvec, wk, svec, prop)
}
\arguments{
\item{A}{matrix}

\item{n}{number of repetition in line}

\item{m}{number of repetition in column}
}
\description{
Internal C++ algorithm for computing the cost matrix.
}
