% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_Gmixt_simultanee.R
\name{Gmixt_simultanee}
\alias{Gmixt_simultanee}
\title{Gmixt_simultanee calculates the cost matrix 
for a segmentation/clustering model}
\usage{
Gmixt_simultanee(Don, lmin, phi)
}
\arguments{
\item{Don}{the bivariate  signal}

\item{lmin}{the minimum size for a segment}

\item{phi}{the  parameters of the mixture}
}
\value{
a matrix
}
\description{
Return matrix G(i,j), the mixture density 
for segment between points (i+1) to j :
\deqn{G(i,j) = \sum_{p=1}^P \log (\pi_p f(y^{ij};\theta_p))}{G(i,j) = \sum{p=1 -> P} log( \\pi{p} f( y{ij} ; \\theta{p} ) ) }
Rq: this density if factorized in order to avoid numerical zeros in
the log
}
