% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gmixt_simultanee_fullcpp}
\alias{Gmixt_simultanee_fullcpp}
\title{Gmixt_simultanee_fullcpp}
\usage{
Gmixt_simultanee_fullcpp(Don, lmin, prop, mu, s)
}
\arguments{
\item{Don}{Bivariate Signal}

\item{lmin}{minimum length of segments}

\item{prop}{mixture parameters}

\item{mu}{mean parameters}

\item{s}{standard deviation paramaters}
}
\description{
C++ function replacing \link{Gmixt_simultanee}
}
