% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_initialisePhi.R
\name{initialisePhi}
\alias{initialisePhi}
\title{initialisePhi is the constructor for a set of parameters for a segclust model}
\usage{
initialisePhi(P, val = -Inf)
}
\arguments{
\item{P}{number of classes}

\item{val}{the value used for initilisation default is -Inf}
}
\value{
a set of parameter phi
}
\description{
initialisePhi is the constructor for a set of parameters for a segclust model
}
