% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_generic.R
\name{likelihood}
\alias{likelihood}
\title{Generic function for likelihood}
\usage{
likelihood(x, ...)
}
\arguments{
\item{x}{object from which likelihood can be extracted}

\item{...}{additional parameters}
}
\description{
Generic function for likelihood
}
