% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_logdens_simultanee.R
\name{logdens_simultanee}
\alias{logdens_simultanee}
\title{logdens_simultanee calculates Gaussian log-densities for a bivariate signal under a mixture model with P components}
\usage{
logdens_simultanee(xk, phi)
}
\arguments{
\item{xk}{the bivaraiate signal}

\item{phi}{parameters of the mixture, P components}
}
\value{
the value of the log density
}
\description{
logdens_simultanee calculates Gaussian log-densities for a bivariate signal under a mixture model with P components
}

