% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_map_plot.R
\name{map_segm}
\alias{map_segm}
\title{Plot map of segmentation}
\usage{
map_segm(data, output, interactive = F, x_col = "expectTime", html = F,
  scale = 100,
  UTMstring = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  width = 400, height = 400, order = NULL)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{position_width}{width between different model if several models are compared}
}
\value{
a graph
}
\description{
\code{plot_segm} plot segmented traject on a map.
}
\examples{
stat_segm(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
