% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_plot_simultanee.R
\name{matrixRupt}
\alias{matrixRupt}
\title{matrixRupt transforms a vector of change point into a data.frame with start
and end of every segment}
\usage{
matrixRupt(x, vectorRupt)
}
\arguments{
\item{x}{the}

\item{vectorRupt}{the vector containing the change point}
}
\value{
the matrix of change point
}
\description{
matrixRupt transforms a vector of change point into a data.frame with start
and end of every segment
}
