% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_mBic.R
\name{mBic}
\alias{mBic}
\title{mBic computes the modified Bic criterion}
\usage{
mBic(Don, resHSimultanee)
}
\arguments{
\item{Don}{the bivariate signal}

\item{resHSimultanee}{the results of hybrid_simultanee}
}
\value{
the mBic criterion for each number of change points K
}
\description{
mBic computes the modified Bic criterion
}
