% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_Mstep_simultanee.R, R/segTraj_Mstep_simultaneeBis.R
\name{Mstep_simultanee}
\alias{Mstep_simultanee}
\title{Mstep_simultanee computes the MLE within the EM framework}
\usage{
Mstep_simultanee(x, rupt, tau, phi, sameSigma = TRUE)

Mstep_simultanee(x, rupt, tau, phi, sameSigma = TRUE)
}
\arguments{
\item{x}{the bivariate signal}

\item{rupt}{the rupture dataframe}

\item{tau}{the K*P matrix containing posterior probabilities of membership to clusters}

\item{phi}{the parameters of the mixture}

\item{x}{the bivariate signal}

\item{rupt}{the rupture dataframe}

\item{phi}{the parameters of the mixture}

\item{tau}{the K*P matrix containing posterior probabilities of membership to clusters}
}
\value{
phi the updated value of the pramaeters

phi the updated value of the pramaeters
}
\description{
Mstep_simultanee computes the MLE within the EM framework

Mstep_simultanee computes the MLE within the EM framework
}

