% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegTraj_EM_cpp.R
\name{Mstep_simultanee_cpp}
\alias{Mstep_simultanee_cpp}
\title{Mstep_simultanee computes the MLE within the EM framework}
\usage{
Mstep_simultanee_cpp(x, rupt, tau, phi, sameSigma = TRUE)
}
\arguments{
\item{x}{the bivariate signal}

\item{rupt}{the rupture dataframe}

\item{tau}{the K*P matrix containing posterior probabilities of membership to clusters}

\item{phi}{the parameters of the mixture}

\item{sameSigma}{whether segments have the same variance}
}
\value{
phi the updated value of the pramaeters
}
\description{
Mstep_simultanee computes the MLE within the EM framework
}
