% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_neighborster.R
\name{neighborster}
\alias{neighborster}
\title{neighbors tests whether neighbors of point k,P can  be used to re-initialize the EM algorithm and to improve the log-likelihood.}
\usage{
neighborster(kv.hull, x, L, k, param, P, lmin, eps, sameSigma)
}
\arguments{
\item{x}{the initial dataset}

\item{L}{the likelihood}

\item{k}{the points of interest}

\item{P}{the number of class}

\item{lmin}{minimal size of the segment to be implemented}
}
\value{
smoothin likelihood
}
\description{
neighbors tests whether neighbors of point k,P can  be used to re-initialize the EM algorithm and to improve the log-likelihood.
}
