% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_series_plot.R
\name{plot_segm}
\alias{plot_segm}
\title{Plot segmentation on time-serie}
\usage{
plot_segm(data, output, separate = T, interactive = F, diag.var,
  x_col = "expectTime", html = F, order = F, stationarity = NULL,
  mean = NULL, var = NULL)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{position_width}{width between different model if several models are compared}
}
\value{
a graph
}
\description{
\code{plot_segm} plot segmented time serie.
}
\examples{
stat_segm(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
