% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmixS4.R
\name{prep_segm_depmixS4}
\alias{prep_segm_depmixS4}
\title{Internal function for HMM}
\usage{
prep_segm_depmixS4(depmixS4.model)
}
\arguments{
\item{depmixS4.model}{depmixS4.model}
}
\description{
\code{prep_segm_depmixS4}
}
