% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{prep_segm_HMM}
\alias{prep_segm_HMM}
\title{Find segment and states for a HMM model}
\usage{
prep_segm_HMM(data, hmm.model)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{hmm.model}{the moveHMM::fitHMM output}
}
\value{
a data.frame with states of the different segments
}
\description{
\code{prep_segm_HMM} find the different segment and states of a given HMM
model
}
\examples{
prep_segm_HMM(data,hmm.model=mod1.hmm)
}
