% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{prep_segm_picard}
\alias{prep_segm_picard}
\title{Find segment and states for a Picard model}
\usage{
prep_segm_picard(data, picard.param, picard.type = "hybrid",
  picard.nseg = NULL)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{picard.param}{the param output of the segmentation}

\item{picard.type}{either 'hybrid' or 'dynprog'}

\item{picard.nseg}{number of segment chosen}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}

\item{type}{type of model either 'hmm' or 'picard'}

\item{hmm.model}{the moveHMM::fitHMM output}
}
\value{
a data.frame with states of the different segments
}
\description{
\code{prep_segm_picard} find the different segment and states of a given HMM
model
}
\examples{
prep_segm_picard(data,picard.param,picard.type='hybrid',picard.nseg=NULL)
}
