% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiftfit.R
\name{prep_segm_shiftfit}
\alias{prep_segm_shiftfit}
\title{Internal function for HMM}
\usage{
prep_segm_shiftfit(data, shiftfit.model)
}
\arguments{
\item{data}{data}

\item{shiftfit.model}{shiftfit.model}
}
\description{
\code{prep_segm_shiftfit}
}
