% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmixS4.R
\name{prepare_depmixS4}
\alias{prepare_depmixS4}
\title{Prepare depmixS4 output for proper plots}
\usage{
prepare_depmixS4(data, depmixS4.model = NULL, diag.var,
  order.var = diag.var[1])
}
\arguments{
\item{data}{data}

\item{depmixS4.model}{depmixS4.model}

\item{diag.var}{diag.var}

\item{order.var}{order.var}
}
\description{
\code{prepare_depmixS4}
}
