% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_HMM.R
\name{prepare_HMM}
\alias{prepare_HMM}
\title{Prepare HMM output for proper comparison plots}
\usage{
prepare_HMM(data, hmm.model = NULL, diag.var, order.var = diag.var[1])
}
\description{
\code{prepare_HMM}
}

