% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_repmat.R
\name{repmat}
\alias{repmat}
\title{repmap repeats a matrix}
\usage{
repmat(a, n, m)
}
\arguments{
\item{a}{the base matrix}

\item{n}{number of repetition in lines}

\item{m}{number of repetition in columns}
}
\value{
a matrix wth n repeats of a in lines et m in columns
}
\description{
repmap repeats a matrix
}

