% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_ruptAsMat.R
\name{ruptAsMat}
\alias{ruptAsMat}
\title{ruptAsMat is an internal function to transform a vector ginvig the change point to matrix 2 columns matrix in whiech each line gives the beginning and the end of a segment}
\usage{
ruptAsMat(vectRupt)
}
\arguments{
\item{vectRupt}{the vector of change point}
}
\value{
the matric containing the semgments
}
\description{
ruptAsMat is an internal function to transform a vector ginvig the change point to matrix 2 columns matrix in whiech each line gives the beginning and the end of a segment
}
