% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segclust2d.R
\docType{package}
\name{segclust2d}
\alias{segclust2d}
\alias{segclust2d-package}
\title{segclust2d: tools for segmentation of animal GPS movement data}
\description{
segclust2d provides usefull tools for the behavioural segmentation of animal
movement
}
