% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segclust.R
\name{segclust_internal}
\alias{segclust_internal}
\title{Internal segmentation/clustering function}
\usage{
segclust_internal(x, seg.var = NULL, diag.var = NULL, order.var = NULL,
  scale.variable, Kmax, ncluster = NULL, lmin = NULL, dat = NULL,
  S = 0.75, type = NULL, sameSigma = F, lissage = F,
  subsample_over = 1000)
}
\description{
Internal segmentation/clustering function
}
