% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_map_plot.R
\name{segmap_list}
\alias{segmap_list}
\title{\code{segmap_list} create maps with a list of object of \code{segmentation}
class}
\usage{
segmap_list(x_list, ncluster_list = NULL, nseg_list = NULL,
  pointsize = 1, linesize = 0.5, coord.names = c("x", "y"))
}
\arguments{
\item{x_list}{list of segmentation objects for different individuals or path}

\item{ncluster_list}{list of number of cluster to be selected for each
individual. If empty, the function takes the default one}

\item{nseg_list}{list of number of segment to be selected for each
individual. If empty, the function takes the default one}

\item{pointsize}{size of points}

\item{linesize}{size of lines}

\item{coord.names}{names of coordinates}
}
\value{
a ggplot2 graph
}
\description{
\code{segmap_list} create maps with a list of object of \code{segmentation}
class
}
