% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_class.R
\name{segmentation-class}
\alias{segmentation-class}
\alias{print.segmentation}
\alias{plot.segmentation}
\alias{likelihood.segmentation}
\alias{stateplot}
\alias{states}
\alias{segment}
\alias{augment.segmentation}
\alias{segmap}
\title{segmentation class description}
\usage{
\method{print}{segmentation}(x, max.level = 1)

\method{plot}{segmentation}(x, nseg = NULL, nclass = NULL, separate = T,
  interactive = F, xcol = "indice", html = F)

\method{likelihood}{segmentation}(x)

stateplot(x, nseg = NULL, nclass = NULL)

states(x, nseg = NULL, nclass = NULL)

segment(x, nseg = NULL, nclass = NULL)

\method{augment}{segmentation}(x, nseg = NULL, nclass = NULL,
  colname_state = "state")

segmap(x, interactive = F, nseg = NULL, nclass = NULL,
  xcol = "expectTime", html = F, scale = 100,
  UTMstring = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  width = 400, height = 400)
}
\arguments{
\item{x}{a \code{segmentation} object generated by
\code{\link{segmentation}}}

\item{nseg}{number of segment chosen}

\item{nclass}{number of classes chosen}

\item{separate}{whether plots should be merged into one graph with several
axis or not (not working yet)}

\item{interactive}{whether plot are interactive (dygraphs/leaflet) or not (ggplot2)}

\item{xcol}{column for x axis. can be POSIXct}

\item{html}{whether htmltools::tagList should be applied on the returned
object for integrating in html pages}

\item{colname_state}{column name for the added state column}

\item{scale}{for dividing coordinates so that it works with leaflet
(interpreting latitude/longitude)}

\item{UTMstring}{projection string for the coordinates.}

\item{width}{for leaflet map}

\item{height}{for leaflet map}
}
\description{
segmentation class description

\code{print.segmentation} prints object of \code{segmentation} class

\code{plot.segmentation} plot object of \code{segmentation} class
- wrapper for \code{\link{plot_segm}}

\code{likelihood.segmentation} plot likelihood estimates of a \code{segmentation} object
- works only for picard segmentation.

\code{stateplot} plot state distribution of a \code{segmentation} object

\code{states} return data.frame with states statistics a \code{segmentation} object

\code{segment} return data.frame with segment information of a \code{segmentation} object

\code{augment.segmentation} return data.frame with original data and state information of a \code{segmentation} object

\code{segmap} create maps with object of \code{segmentation} class
}
