% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_hmm}
\alias{segmentation_hmm}
\title{Segmentation Function for HMM
Wrapper for moveHMM::fitHMM}
\usage{
segmentation_hmm(data, nbStates, mu0, sigma0, zeromass0, angleMean0, kappa0,
  stepDist = "gamma", angleDist = "vm", coordNames = c("x", "y"),
  coord.type = c("UTM"), timecol = "expectTime")
}
\arguments{
\item{data}{data.frame}

\item{nbStates}{number of states fitted}

\item{mu0}{a priori mean of dist}

\item{sigma0}{a priori sd of dist}

\item{zeromass0}{a priori zeromass of dist}

\item{angleMean0}{a priori mean of angle}

\item{kappa0}{a priori kappa of angle}

\item{angleDist}{type of HMM}

\item{coordNames}{names of coordinates in data}

\item{coord.type}{whether coordinates are in UTM or lat/lon}

\item{timecol}{column names for time in data}

\item{hmm.param}{parameters used for fitting the hmm model}
}
\value{
a segmentation object
}
\description{
Segmentation Function for HMM
Wrapper for moveHMM::fitHMM
}
