% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_internal}
\alias{segmentation_internal}
\title{Internal segmentation function}
\usage{
segmentation_internal(
  x,
  seg.var,
  diag.var,
  order.var,
  lmin,
  Kmax,
  scale.variable,
  sameSigma = FALSE,
  ...
)
}
\arguments{
\item{x}{data.frame with observations}

\item{seg.var}{for behavioral segmentation: names of the variables used for
segmentation (either one or two names).}

\item{diag.var}{names of the variables on which
statistics are calculated.}

\item{order.var}{names of the variable with which states are ordered.}

\item{lmin}{minimum length of segments.}

\item{Kmax}{maximum number of segments.}

\item{scale.variable}{TRUE or FALSE for automatic scaling of variables
(reduction and  centering)}

\item{sameSigma}{does segments have same variance ?}

\item{...}{additional parameters given to \link{chooseseg_lavielle}}
}
\description{
Internal segmentation function
}
