% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_picard}
\alias{segmentation_picard}
\title{Segmentation Function for Picard/segTraj}
\usage{
segmentation_picard(data, seg.var = NULL, diag.var = NULL,
  order.var = NULL, picard.type = "DynProg", scale.variable = F,
  nclass = NULL, Kmax = NULL, lmin = NULL, sameSigma = F)
}
\arguments{
\item{picard.type}{either 'hybrid' or 'dynprog'}

\item{scale.variable}{if variable needs to be scaled for segmentation}

\item{nclass}{number of class for hybrid_simultanee}

\item{Kmax}{maximum number of segments}

\item{lmin}{minimum size of segments}

\item{sameSigma}{whether segment should have equal variance or not.}
}
\description{
Segmentation Function for Picard/segTraj
}
