% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtools.R
\docType{package}
\name{segtools}
\alias{segtools}
\alias{segtools-package}
\title{segtools: tools for segmentation of animal GPS movement data}
\description{
segtools provides usefull tools for the behavioural segmentation of animal
movement
}
