% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_data.R
\docType{data}
\name{simulmode}
\alias{simulmode}
\title{Simulations of behavioural mode}
\format{
A data frame with 302 rows and 10 variables:
\describe{
  \item{indice}{index of position}
  \item{x}{x coordinates}
  \item{y}{y coordinates}
  \item{speed}{smoothed speed}
  \item{spatial_angle}{angle at constant step length}
  \item{dist}{raw speed}
  \item{angle}{angular speed}
  \item{vit_p}{persistence speed}
  \item{vit_r}{rotation speed}
  \item{vit_p_spa}{persistence speed calculated with spatial angles}
  \item{vit_r_spa}{rotation speed calculated with spatial angles}
  \item{dateTime}{arbitrary date in POSIXct format}
}
}
\usage{
simulmode
}
\description{
A dataset containing a simulation of 3 different behavioural mode
}
\keyword{datasets}
