% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_data.R
\docType{data}
\name{simulshift}
\alias{simulshift}
\title{Simulations of home-range shift}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{indice}{index of position}
  \item{x}{x coordinates}
  \item{y}{y coordinates}
  \item{dateTime}{arbitrary date in POSIXct format}
}
}
\usage{
simulshift
}
\description{
A dataset containing a simulation of home-range shift
}
\keyword{datasets}
