% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{stat_segm}
\alias{stat_segm}
\title{Calculate statistics on a given segmentation}
\usage{
stat_segm(data, diag.var, order.var = NULL, model.type = "hmm",
  hmm.model = NULL, picard.param = NULL, picard.type = NULL, picard.nseg)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}

\item{hmm.model}{the moveHMM::fitHMM output}

\item{picard.param}{the param output of the segmentation}

\item{picard.type}{either 'hybrid' or 'dynprog'}

\item{picard.nseg}{number of segment chosen}

\item{type}{type of model either 'hmm' or 'picard'}
}
\value{
a list which first element is a data.frame with states of the
  different segments and which second element is a data.frame with mean and
  variance of the different states
}
\description{
\code{stat_segm} calculates statistics of a given segmentation : mean and
variance of the different states. it also creates standard objects for plot.
Actually implementation exists for \code{moveHMM} and \code{segTraj} methods.
}
\examples{
stat_segm(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
