% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmixS4.R
\name{stat_segm_depmixS4}
\alias{stat_segm_depmixS4}
\title{Get segment statistic for HMM model}
\usage{
stat_segm_depmixS4(data, depmixS4.model = NULL, diag.var, order.var = NULL)
}
\description{
\code{stat_segm_depmixS4}
}
