% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{subsample_rename}
\alias{subsample_rename}
\title{Internal function for subsampling}
\usage{
subsample_rename(df, fulldata, colname)
}
\arguments{
\item{df}{subsampled data.frame with additional information on segmentation}

\item{fulldata}{full data.frame}

\item{colname}{column name}
}
\description{
merge subsampled data.frame df with fulldata to add segmentation information
on the full data.frame
}
