% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_segment.R
\name{test_stationarity}
\alias{test_stationarity}
\title{test_stationarity checks the stationarity within each segment}
\usage{
test_stationarity(dat, outputs, Kmax)
}
\description{
\code{test_stationarity}
}
