% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_DynProg.R
\name{wrap_dynprog_cpp}
\alias{wrap_dynprog_cpp}
\title{DynProg RCpp
DynProg computes the change points given a cost matrix matD
and a maximum number of segments Kmax}
\usage{
wrap_dynprog_cpp(G, K)
}
\arguments{
\item{G}{the cost Matrix os size n x n}

\item{K}{the number of segments considered}
}
\value{
a list with J.est a vector with Kmax value, the Kth is the minimum
  contrast for a model with K segments (-J.est is the log-likelihood) and
  with  t.test a matrix, line K are the coordinates of the change points for
  a model with K segments
}
\description{
DynProg RCpp
DynProg computes the change points given a cost matrix matD
and a maximum number of segments Kmax
}
