% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{angular_speed}
\alias{angular_speed}
\title{Calculate angular speed along a path}
\usage{
angular_speed(x, coord.names = c("x", "y"))
}
\arguments{
\item{x}{data.frame with locations}

\item{coord.names}{names of coordinates column in \code{x}}
}
\value{
vector of turning angle.
}
\description{
\code{angular_speed} calculate turning angle between locations, taking a
dataframe as input.
}
\author{
Remi Patin, Simon Benhamou.
}
