% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Kmax.R
\name{chooseseg_lavielle}
\alias{chooseseg_lavielle}
\title{Internal function for choosing optimal number of segment}
\usage{
chooseseg_lavielle(J.est, S = 0.75)
}
\arguments{
\item{J.est}{data used for segmentation. Supported : data.frame, 2-columns matrix, move object, ltraj object}

\item{S}{threshold for choosing the number. See adehabitatLT::chooseseg}
}
\value{
a list with optimal number of segment and full data.frame of the calculus
}
\description{
Internal function for choosing optimal number of segment using lavielle method
modified frome adehabitatLT::chooseseg. Method based on identifying breaks in the slope of the contrast
}
\examples{
segmentation(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}

