% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_EM.init_simultanee.R
\name{EM.init_simultanee}
\alias{EM.init_simultanee}
\title{EM.init_simultanee proposes an inial value for the EM algorithm  based on a hierarchical clustering  algorithm (ascending)}
\usage{
EM.init_simultanee(x, rupt, K, P)
}
\arguments{
\item{x}{the bivariate signal}

\item{rupt}{the  change point instants, data.frame}
}
\description{
EM.init_simultanee proposes an inial value for the EM algorithm  based on a hierarchical clustering  algorithm (ascending)
}
