% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{find_mu_sd}
\alias{find_mu_sd}
\title{Find mean and standard deviation of segments}
\usage{
find_mu_sd(df.states, diag.var)
}
\arguments{
\item{df.states}{a list of data.frame}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}

\item{df.segm}{output of prep_segm function}
}
\value{
a data.frame with mean and variance of the different states
}
\description{
\code{find_mu_sd} calculates statistics of a given segmentation : mean
and variance of the different states.
}
\examples{
calc_stat_states(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}

