% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gmixt_algo_cpp}
\alias{Gmixt_algo_cpp}
\title{Gmixt_algo_cpp}
\usage{
Gmixt_algo_cpp(zi, lgi, P, mvec, wk, svec, prop)
}
\arguments{
\item{zi}{vector of observations}

\item{lgi}{vector of indices}

\item{P}{number of class}

\item{mvec}{vector of means for each class}

\item{wk}{temporary vector for calculations}

\item{svec}{vector of standard deviations for each class}

\item{prop}{mixture vector}
}
\description{
Internal C++ algorithm for computing the cost matrix.
}
