% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_initialisePhi.R
\name{initialisePhi}
\alias{initialisePhi}
\title{initialisePhi is the constructor for a set of parameters for a segclust model
P number of classes
val the value ti be used for tinitiliastation default is -Inf}
\usage{
initialisePhi(P, val = -Inf)
}
\value{
a set of paraemter phi
}
\description{
initialisePhi is the constructor for a set of parameters for a segclust model
P number of classes
val the value ti be used for tinitiliastation default is -Inf
}

