% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_map_plot.R
\name{map_segm}
\alias{map_segm}
\title{Plot map of segmentation}
\usage{
map_segm(data, output, interactive = F, html = F, scale = 100,
  UTMstring = "+proj=longlat +datum=WGS84 +no_defs", width = 400,
  height = 400, order = NULL, pointsize = 1, linesize = 0.5)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{output}{outputs of the segmentation  or segclust algorithm for one
number of segment}

\item{interactive}{should graph be interactive through leaflet ?}

\item{html}{should the graph be incorporated in a markdown file through
htmltools::tagList()}

\item{scale}{scaling paremeters to represent data with leaflet}

\item{UTMstring}{projection of the coordinates}

\item{width}{width}

\item{height}{height}

\item{order}{should cluster be ordered}

\item{pointsize}{size of points}

\item{linesize}{size of lines}
}
\value{
a graph
}
\description{
\code{plot_segm} plot segmented traject on a map.
}
\examples{
\dontrun{stat_segm(data, diag.var=c("dist","angle"),
order.var='dist',type='hmm',hmm.model=mod1.hmm)}
}
