% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_series_plot.R
\name{plot_segm}
\alias{plot_segm}
\title{Plot segmentation on time-serie}
\usage{
plot_segm(data, output, interactive = F, diag.var, x_col = "expectTime",
  html = F, order = F)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{output}{outputs of the segmentation  or segclust algorithm for one number of segment}

\item{interactive}{should graph be interactive through leaflet ?}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{x_col}{column name for time}

\item{html}{should the graph be incorporated in a markdown file through htmltools::tagList()}

\item{order}{should cluster be ordered}
}
\value{
a graph
}
\description{
\code{plot_segm} plot segmented time serie.
}
