% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_states_plot.R
\name{plot_states}
\alias{plot_states}
\title{Plot states statistics}
\usage{
plot_states(outputs, diag.var, position_width = 0.3)
}
\arguments{
\item{diag.var}{names of the variables on which statistics are calculated}

\item{position_width}{width between different model if several models are compared}

\item{data}{the data.frame with the different variable}
}
\value{
a graph
}
\description{
\code{plot_states} plot states statistics
}
\examples{
stat_segm(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
