% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare.R
\name{prep_segm}
\alias{prep_segm}
\title{Find segment and states for a Picard model}
\usage{
prep_segm(data, param, seg.type = NULL, nseg = NULL)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{param}{the param output of the segmentation}

\item{seg.type}{either 'hybrid' or 'dynprog'}

\item{nseg}{number of segment chosen}
}
\value{
a data.frame with states of the different segments
}
\description{
\code{prep_segm} find the different segment and states of a given HMM
model
}
