% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiftfit.R
\name{prepare_shiftfit}
\alias{prepare_shiftfit}
\title{Prepare shiftfit output for proper comparison plots}
\usage{
prepare_shiftfit(data, shiftfit.model = NULL, diag.var,
  order.var = diag.var[1])
}
\description{
\code{prepare_shiftfit}
}

