% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segTraj_ruptAsMat.R
\name{ruptAsMat}
\alias{ruptAsMat}
\title{ruptAsMat is an internal function to transform a vector giving the change
point to matrix 2 columns matrix in which each line gives the beginning and
the end of a segment}
\usage{
ruptAsMat(vectRupt)
}
\arguments{
\item{vectRupt}{the vector of change point}
}
\value{
the matrix containing the segments
}
\description{
ruptAsMat is an internal function to transform a vector giving the change
point to matrix 2 columns matrix in which each line gives the beginning and
the end of a segment
}
