% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segclust.R
\name{segclust}
\alias{segclust}
\alias{segclust.data.frame}
\alias{segclust.Move}
\alias{segclust.ltraj}
\title{Segmentation/Clustering of movement data - Generic function}
\usage{
segclust(x, ...)

\method{segclust}{data.frame}(x, Kmax, lmin, ncluster, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("x", "y"), ...)

\method{segclust}{Move}(x, Kmax, lmin, ncluster, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("x", "y"), ...)

\method{segclust}{ltraj}(x, Kmax, lmin, ncluster, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("x", "y"), ...)
}
\arguments{
\item{x}{data used for segmentation. Supported : data.frame, 2-columns matrix, move object, ltraj object}

\item{...}{additional parameters}

\item{Kmax}{maximum number of segments. Default to 10.}

\item{lmin}{minimum size of segments. Default to length of time series/Kmax/2.}

\item{ncluster}{number or list of cluster to be chosen.}

\item{type}{type of segmentation. Either "home-range" or "behavior"}

\item{seg.var}{for behavioral segmentation : names of the variables used for segmentation (either one or two names)}

\item{diag.var}{for behavioral segmentation : names of the variables on which statistics are calculated}

\item{order.var}{for behavioral segmentation : names of the variable with which states are ordered}

\item{coord.names}{for home-range segmentation and data.frame, names of coordinates. Default x and y. Not needed for move and ltraj objects}
}
\value{
a \code{\link{segmentation-class}} object
}
\description{
Joint Segmentation/Clustering of movement data. Method available for data.frame, move and ltraj object

Segmentation/Clustering function for data.frame

Segmentation/Clustering function for move objects

Segmentation/Clustering function for ltraj objects
}
\examples{
\dontrun{segmentation(data, diag.var = c("dist","angle"),
order.var='dist',type='hmm',hmm.model=mod1.hmm)}
}
