% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segclust.R
\name{segclust_internal}
\alias{segclust_internal}
\title{Internal segmentation/clustering function}
\usage{
segclust_internal(x, seg.var = NULL, diag.var = NULL, order.var = NULL,
  scale.variable = scale.variable, Kmax = NULL, ncluster = NULL,
  lmin = NULL, dat = NULL, data.type = NULL, S = NULL, type = NULL,
  sameSigma = NULL)
}
\description{
Internal segmentation/clustering function
}
