% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segclust.R
\name{segclust_internal}
\alias{segclust_internal}
\title{Internal segmentation/clustering function}
\usage{
segclust_internal(
  x,
  seg.var,
  diag.var,
  order.var,
  Kmax,
  ncluster,
  lmin,
  scale.variable,
  sameSigma = FALSE,
  ...
)
}
\arguments{
\item{x}{data used for segmentation. Supported: data.frame, Move object,
ltraj object}

\item{seg.var}{for behavioral segmentation: names of the variables used for
segmentation (either one or two names).}

\item{diag.var}{names of the variables on which
statistics are calculated.}

\item{order.var}{names of the variable with which states are ordered.}

\item{Kmax}{maximum number of segments.}

\item{ncluster}{number of cluster into which segments should be grouped. Can
be a vector if one want to test several number of clusters.}

\item{lmin}{minimum length of segments.}

\item{scale.variable}{TRUE or FALSE for automatic scaling of variables
(reduction and  centering)}

\item{sameSigma}{does segments have same variance ?}

\item{...}{additional arguments given to \code{\link{chooseseg_lavielle}}}
}
\description{
Internal segmentation/clustering function
}
