% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\alias{segmentation.data.frame}
\alias{segmentation.Move}
\alias{segmentation.ltraj}
\title{Segmentation of movement data - Generic function}
\usage{
segmentation(x, ...)

\method{segmentation}{data.frame}(x, Kmax, lmin, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("x", "y"), ...)

\method{segmentation}{Move}(x, Kmax, lmin, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("coords.x1", "coords.x2"), ...)

\method{segmentation}{ltraj}(x, Kmax, lmin, type = "home-range",
  seg.var = NULL, diag.var = seg.var, order.var = seg.var[1],
  coord.names = c("x", "y"), ...)
}
\arguments{
\item{x}{data used for segmentation. Supported : data.frame, 2-columns matrix, move object, ltraj object}

\item{Kmax}{maximum number of segments. Default to 10.}

\item{lmin}{minimum size of segments. Default to length of time series/Kmax/2.}

\item{type}{type of segmentation. Either "home-range" or "behavior"}

\item{seg.var}{for behavioral segmentation : names of the variables used for segmentation (either one or two names)}

\item{diag.var}{for behavioral segmentation : names of the variables on which statistics are calculated}

\item{order.var}{for behavioral segmentation : names of the variable with which states are ordered}

\item{coord.names}{for home-range segmentation and data.frame, names of coordinates. Default x and y. Not needed for move and ltraj objects}

\item{scale.variable}{for behavioral segmentation : whether variables need to be scaled for segmentation}
}
\value{
a \code{\link{segmentation-class}} object
}
\description{
Segmentation of movement data. No clustering. Method available for data.frame,  move and ltraj object

Segmentation function for data.frame

Segmentation function for move objects

Segmentation function for ltraj objects
}
\examples{
segmentation(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
