% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R, R/segmentation_class.R
\name{segmentation}
\alias{segmentation}
\alias{print.segmentation}
\alias{plot.segmentation}
\alias{likelihood.segmentation}
\alias{stateplot}
\alias{states}
\alias{segment}
\alias{augment.segmentation}
\alias{segmap}
\title{Segmentation of movement data}
\usage{
segmentation(data, seg.var, diag.var = seg.var,
  order.var = dplyr::first(diag.var), type = "picard", ...)

\method{print}{segmentation}(x, max.level = 1)

\method{plot}{segmentation}(x, nseg = NULL, separate = T, interactive = F,
  xcol = "indice", html = F)

\method{likelihood}{segmentation}(x)

stateplot(x, nseg = NULL)

states(x, nseg = NULL)

segment(x, nseg = NULL)

\method{augment}{segmentation}(x, nseg = NULL, colname_state = "state")

segmap(x, interactive = F, nseg = NULL, x_col = "expectTime", html = F,
  scale = 100,
  UTMstring = "+proj=utm +zone=35 +south +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  width = 400, height = 400)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{seg.var}{names of the variables used for segmentation (if picard)}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}

\item{type}{type of model either 'hmm' or 'picard'}

\item{...}{other arguments passed to segmentation_HMM and segmentation_picard}
}
\value{
a \code{\link{segmentation}} object
}
\description{
Segmentation of movement data. Actually implementation exists for
\code{moveHMM} and \code{segTraj} methods. The function calls either
\code{\link{segmentation_HMM}} or \code{\link{segmentation_picard}}

Print object of \code{segmentation} class

Plot object of \code{segmentation} class
Wrapper for \code{\link{plot_segm}}

Plot likelihood estimates of a \code{segmentation} object
Works only for picard segmentation.

Plot state distribution of a \code{segmentation} object

Return data.frame with states statistics a \code{segmentation} object

Return data.frame with segment information of a \code{segmentation} object

Return data.frame with original data and state information of a \code{segmentation} object

Create maps with object of \code{segmentation} class
}
\examples{
segmentation(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
