% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\title{Segmentation of movement data}
\usage{
segmentation(data, seg.var, diag.var = seg.var,
  order.var = dplyr::first(diag.var), type = "picard", picard.type = NULL,
  scale.variable = F, ...)
}
\arguments{
\item{data}{the data.frame with the different variable}

\item{seg.var}{names of the variables used for segmentation (either one or two names)}

\item{diag.var}{names of the variables on which statistics are calculated}

\item{order.var}{names of the variable with which states are ordered}

\item{type}{type of model either 'hmm' or 'picard'}

\item{picard.type}{whether mode should be segmentation only (DynProg),
clustering-segmentation with a given number of class (hybrid_simultanee) or
with an unknown number of class (variable_class).}

\item{scale.variable}{whether variables need to be scaled for segmentation}
}
\value{
a \code{\link{segmentation-class}} object
}
\description{
Segmentation of movement data. Actually implementation exists for
\code{moveHMM} and \code{segTraj} methods. The function calls either
\code{\link{segmentation_HMM}} or \code{\link{segmentation_picard}}
}
\examples{
segmentation(data,diag.var=c("dist","angle"),order.var='dist',type='hmm',hmm.model=mod1.hmm)
}
