% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_hmm}
\alias{segmentation_hmm}
\title{Segmentation Function for HMM}
\usage{
segmentation_hmm(data, nbStates, mu0, sigma0, zeromass0, angleMean0, kappa0,
  stepDist = "gamma", angleDist = "vm", coordNames = c("x", "y"),
  coord.type = c("UTM"), timecol = "expectTime")
}
\arguments{
\item{hmm.param}{parameters used for fitting the hmm model}
}
\description{
Segmentation Function for HMM
}
