% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_internal}
\alias{segmentation_internal}
\title{Internal segmentation function}
\usage{
segmentation_internal(x, seg.var = NULL, diag.var = NULL,
  order.var = NULL, scale.variable = scale.variable, Kmax = NULL,
  lmin = NULL, dat = NULL, data.type = NULL, S = NULL, type = NULL,
  sameSigma = NULL)
}
\description{
Internal segmentation function
}

