% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation_internal}
\alias{segmentation_internal}
\title{Internal segmentation function}
\usage{
segmentation_internal(x, seg.var = NULL, diag.var = NULL,
  order.var = NULL, scale.variable = NULL, Kmax, lmin = NULL,
  dat = NULL, S = 0.75, type = NULL, sameSigma = F,
  subsample_over = 10000)
}
\description{
Internal segmentation function
}
