% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulshift.R
\docType{data}
\name{simulshift}
\alias{simulshift}
\title{Simulations of home-range shift}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{indice}{index of position}
  \item{x}{x coordinates}
  \item{y}{y coordinates}
  \item{expectTime}{arbitrary POSIXct time}#'
}}
\usage{
simulshift
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
