% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_covariates.R
\name{spatial_angle}
\alias{spatial_angle}
\title{Calculate spatial angle along a trajectory}
\usage{
spatial_angle(df, coord.names = c("x", "y"), radius = NULL)
}
\arguments{
\item{coord.names}{names of coordinates column in \code{x}}

\item{radius}{for angle calculation. Default is median of step length.}

\item{x}{data.frame with locations}
}
\value{
vector of spatial angle.
}
\description{
\code{spatial_angle} calculate spatial angle between locations, taking a dataframe as input. Spatial angle is considered as the angle between the focus point, the first location entering a given circle and the last location inside.
}
\examples{
calc_speed(df,coord.names = c("x","y"), timecol = "dateTime", smoothed = T)
}
\author{
Rémi Patin, \email{remi.patin@normale.fr}, Simon Benhamou.
}
