% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiftfit.R
\name{stat_segm_shiftfit}
\alias{stat_segm_shiftfit}
\title{Get segment statistic for HMM model}
\usage{
stat_segm_shiftfit(data, shiftfit.model = NULL, diag.var, order.var = NULL)
}
\description{
\code{stat_segm_shiftfit}
}

