% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{subsample}
\alias{subsample}
\title{Internal function for subsampling}
\usage{
subsample(x, subsample_over)
}
\arguments{
\item{x}{data.frame to be subsampled}

\item{subsample_over}{maximum number of row accepted}
}
\description{
if nrow(x) > subsample_over, subsample with the minimum needed to get a
data.frame smaller than subsample_over
}
